Package "discord" is removed. "CommandHandler" and "ReactionsHandler" are moved to "handlers" package. New class - "HandlersKeeper". It is keeping all handlers, exclude ReactionsHandler and CommandHandler Method "isNumber" in "DiceRollHandler" class was renamed to "isPositiveNumber".
