+++ src/main/java/tableBot/games/dice/Thrower.java
@@ -1,5 +1,7 @@
 package tableBot.games.dice;
 
+import org.jetbrains.annotations.NotNull;
+
 import java.util.Arrays;
 import java.util.Random;
 
@@ -7,11 +9,13 @@ public class Thrower
 {
     private static Random random = new Random();
 
+    @NotNull
     public static String throwCoin ()
     {
         return random.nextInt(2) == 0 ?  "obverse" : "reverse";
     }
 
+    @NotNull
     public static String throwDice (int diceCount, int SidesCount)
     {
         int[] result = new int[diceCount];
