+++ src/main/java/tableBot/handlers/CommandHandler.java
@@ -1,26 +1,18 @@
-package tableBot.discord;
+package tableBot.handlers;
 
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.TextChannel;
 import net.dv8tion.jda.api.entities.User;
-import net.dv8tion.jda.api.events.message.guild.react.GuildMessageReactionAddEvent;
-import org.jetbrains.annotations.NotNull;
 
 import tableBot.InfoGetter;
 import tableBot.PathGetter;
-import tableBot.handlers.DiceRollHandler;
-import tableBot.handlers.GallowsHandler;
-import tableBot.handlers.CardGameHandler;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageUpdateEvent;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 
 public class CommandHandler extends ListenerAdapter
 {
-    private final String prefix = "-";
-    private final GallowsHandler gallowsHandler = new GallowsHandler();
-    private final CardGameHandler cardGameHandler = new CardGameHandler();
-    private final DiceRollHandler diceRollHandler = new DiceRollHandler();
+    private final static String prefix = "-";
 
     public void onGuildMessageReceived (GuildMessageReceivedEvent event)
     {
@@ -42,13 +34,13 @@ public class CommandHandler extends ListenerAdapter
                 InfoGetter.showHelp(channel, path);
                 break;
             case (prefix + "gallows"):
-                gallowsHandler.handleCommand(channel, command, user);
+                HandlersKeeper.gallowsHandler.handleCommand(channel, command, user);
                 break;
             case (prefix + "cards"):
-                cardGameHandler.handleCommand(channel, command, user);
+                HandlersKeeper.cardGameHandler.handleCommand(channel, command, user);
                 break;
             case (prefix + "throw"):
-                diceRollHandler.handleCommand(channel, command, user);
+                HandlersKeeper.diceRollHandler.handleCommand(channel, command, user);
                 break;
             case (prefix + "activity"):
                 channel.sendMessage(InfoGetter.showActivities()).queue();
