+++ src/main/java/tableBot/handlers/HandlersKeeper.java
@@ -1,0 +1,8 @@
+package tableBot.handlers;
+
+class HandlersKeeper
+{
+    static GallowsHandler gallowsHandler = new GallowsHandler();
+    static CardGameHandler cardGameHandler = new CardGameHandler();
+    static DiceRollHandler diceRollHandler = new DiceRollHandler();
+}
